Adjust to OTSpanContext changes and cleanse
In this case, "cleanse" involves a bnuch of Pods work to make the example reflect the reality of what a user of this pod will do. Also, wherever possible, LS* has been replaced by OT*. This includes the singleton, which should be an OT thing, not an LS thing.

As with other SpanContext changes, this PR cleans up some subtle cases in extract and leads to better separation of concerns around causal relationships between spans.