Add support for `make build_protobuf`
This ends up being really messy because cocoapods.

More seriously, this sucks because cocoapods won't run arbitrary code
after packages are installed (or at least it won't do so in this
config), and it won't even install pods unless there's some sort of
Xcode project to bind with. Hence PodInstallTarget.